requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
openai>=1.59.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-multipart>=0.0.9
